Include dom macros
The dom macros was not included from clojurescript, which caused the fallback to be run all the time. The fallback function parses using spec/conform which adds a significat amount of time (10x and more) to the renders.

I don't know why it wasn't previously included and if there might be any reason to why it was left out.